refactor(links): shared saas-aware resolveTenantSlug for all public links; fix agreement localhost/empty-slug
Move resolveTenantSlug from server/api/inspections.ts to server/lib/url.ts
alongside getBaseUrl/getBookingHost so it is accessible to all API modules.

Fix all three agreement link build sites in admin.ts (sendAgreementRoute,
getSignerLinkRoute, remindSignerRoute) to use await resolveTenantSlug()
instead of the raw requestedTenantSlug that returns empty string on saas
authenticated routes, producing http://localhost/agreements/sign//<token>.

Fix invoices.ts resolveSignatureInspectorForEmail() the same way so the
inspector booking-link footer in payment-request emails resolves the
correct tenant slug on saas.

Add APP_BASE_URL to wrangler.saas.jsonc (gitignored) so getBookingHost
resolves correctly in every context including scheduled/queue runs.

Update tests/unit/resolve-tenant-slug.spec.ts import path to new location.
